Reorganize folder structure of nodes
