+++ src/unitaria/nodes/basic/adjoint.py
