+++ src/unitaria/nodes/basic/block_horizontal.py
@@ -1,13 +1,12 @@
 import numpy as np
 
 from unitaria.nodes.node import Node
-from unitaria.nodes.basic_operations.scale import Scale
-from unitaria.nodes.basic_operations.adjoint import Adjoint
-from unitaria.nodes.basic_operations.unsafe_multiplication import UnsafeMul
-from unitaria.nodes.basic_operations.tensor import Tensor
-
 from unitaria.nodes.proxy_node import ProxyNode
-from unitaria.nodes.controlled_operations.block_diagonal import BlockDiagonal
+from unitaria.nodes.basic.scale import Scale
+from unitaria.nodes.basic.adjoint import Adjoint
+from unitaria.nodes.basic.unsafe_multiplication import UnsafeMul
+from unitaria.nodes.basic.tensor import Tensor
+from unitaria.nodes.basic.block_diagonal import BlockDiagonal
 from unitaria.nodes.permutation import Permutation
 from unitaria.nodes.constants.constant_vector import ConstantVector
 from unitaria.nodes.identity import Identity
