+++ src/unitaria/nodes/basic/tensor.py
