+++ tests/nodes/test_basic_ops.py
@@ -1,8 +1,8 @@
 from unitaria.nodes.identity import Identity
-from unitaria.nodes.integer_arithmetic.increment import Increment
+from unitaria.nodes.classical.increment import Increment
 from unitaria.subspace import Subspace
 from unitaria.verifier import verify
-from unitaria.nodes.basic_operations.scale import Scale
+from unitaria.nodes.basic.scale import Scale
 
 
 def test_tensor():
