+++ tests/nodes/test_block_concatenation.py
@@ -1,6 +1,6 @@
-from unitaria.nodes.block_concatenation.block_horizontal import BlockHorizontal
-from unitaria.nodes.block_concatenation.block_vertical import BlockVertical
-from unitaria.nodes.integer_arithmetic.increment import Increment
+from unitaria.nodes.basic.block_horizontal import BlockHorizontal
+from unitaria.nodes.basic.block_vertical import BlockVertical
+from unitaria.nodes.classical.increment import Increment
 from unitaria.nodes.identity import Identity
 from unitaria.subspace import Subspace
 from unitaria.verifier import verify
