+++ tests/nodes/test_integer_arithmetic.py
@@ -1,10 +1,10 @@
 import pytest
 
-from unitaria.nodes.integer_arithmetic.increment import Increment
+from unitaria.nodes.classical.increment import Increment
 
-from unitaria.nodes.integer_arithmetic.integer_addition import IntegerAddition
-from unitaria.nodes.integer_arithmetic.constant_integer_addition import ConstantIntegerAddition
-from unitaria.nodes.integer_arithmetic.constant_integer_multiplication import ConstantIntegerMultiplication
+from unitaria.nodes.classical.integer_addition import IntegerAddition
+from unitaria.nodes.classical.constant_integer_addition import ConstantIntegerAddition
+from unitaria.nodes.classical.constant_integer_multiplication import ConstantIntegerMultiplication
 
 from unitaria.verifier import verify
