+++ tests/nodes/test_ring.py
@@ -1,4 +1,4 @@
-from unitaria.nodes.integer_arithmetic.increment import Increment
+from unitaria.nodes.classical.increment import Increment
 from unitaria.nodes.identity import Identity
 from unitaria.subspace import Subspace
 from unitaria.verifier import verify
